tests: moved usevariable to its own spec file
