+++ test/core/useQuery.spec.js
